+++ packages/liveblocks-react-ui/src/shared.ts
@@ -1,13 +1,13 @@
 import type { BaseUserMeta, Client } from "@liveblocks/core";
 import { kInternal, raise, stringify } from "@liveblocks/core";
 import {
+  ClientContext,
+  RoomContext,
   useClient,
-  useLiveblocksContextBundleOrNull__,
   useRoom,
-  useRoomContextBundleOrNull__,
   useSelf,
 } from "@liveblocks/react";
-import React from "react";
+import React, { useContext, useSyncExternalStore } from "react";
 
 type OpaqueClient = Client<BaseUserMeta>;
 
@@ -104,14 +104,22 @@ function useCurrentUserIdFromRoom() {
   return useSelf((user) => (typeof user.id === "string" ? user.id : null));
 }
 
-export function useCurrentUserId(): string | null {
-  const roomContextBundle = useRoomContextBundleOrNull__();
-  const liveblocksContextBundle = useLiveblocksContextBundleOrNull__();
+function useCurrentUserIdFromClient_withClient(client: OpaqueClient) {
+  const currentUserIdStore = client[kInternal].currentUserIdStore;
+  return useSyncExternalStore(
+    currentUserIdStore.subscribe,
+    currentUserIdStore.get,
+    currentUserIdStore.get
+  );
+}
 
-  if (roomContextBundle !== null) {
+export function useCurrentUserId(): string | null {
+  const client = useContext(ClientContext);
+  const room = useContext(RoomContext);
+  if (room !== null) {
     return useCurrentUserIdFromRoom();
-  } else if (liveblocksContextBundle !== null) {
-    return liveblocksContextBundle[kInternal].useCurrentUserIdFromClient();
+  } else if (client !== null) {
+    return useCurrentUserIdFromClient_withClient(client);
   } else {
     raise(
       "LiveblocksProvider or RoomProvider are missing from the React tree."
